+++ l10n/lv.js
@@ -8,9 +8,9 @@ OC.L10N.register(
     "You can close this page. You will be notified when it finishes." : "Šo lapu var aizvērt. Tiks paziņot, kad tas tiks pabeigts.",
     "No Google OAuth app configured. Ask your Nextcloud administrator to configure Google connected accounts admin section." : "Kav konfigurēta neviena Google OAuth lietotne. Jāvaicā savam Nextcloud pārvaldītājam, lai konfigurē sasaistīto google kontu pārvaldības sadaļu.",
     "Authentication" : "Autentifikācija",
+    "Connected as {user}" : "Savienojies kā {user}",
     "Contacts" : "Kontakti",
     "Import calendar" : "Importēt kalendāru",
-    "Photos" : "Bildes",
     "Drive" : "Braukt"
 },
 "nplurals=3; plural=(n%10==1 && n%100!=11 ? 0 : n != 0 ? 1 : 2);");