+++ l10n/pt_PT.js
@@ -8,7 +8,6 @@ OC.L10N.register(
     "Authentication" : "Autenticação",
     "Contacts" : "Contactos",
     "Calendars" : "Calendários",
-    "Import calendar" : "Importar calendário",
-    "Photos" : "Fotos"
+    "Import calendar" : "Importar calendário"
 },
 "nplurals=3; plural=(n == 0 || n == 1) ? 0 : n != 0 && n % 1000000 == 0 ? 1 : 2;");