+++ l10n/uz.js
@@ -2,10 +2,10 @@ OC.L10N.register(
     "integration_google",
     {
     "Bad credentials" : "Akkaunt ma'lumotlari xato",
+    "Connected accounts" : "Ulangan akkauntlar",
     "Client ID" : "Mijoz identifikatori",
     "Authentication" : "Autentifikatsiya",
     "Contacts" : "Contacts",
-    "Calendars" : "Taqvimlar",
-    "Photos" : "Photos"
+    "Calendars" : "Taqvimlar"
 },
 "nplurals=1; plural=0;");