+++ l10n/vi.js
@@ -7,7 +7,6 @@ OC.L10N.register(
     "Contacts" : "Liên hệ",
     "Calendars" : "Lịch",
     "Import calendar" : "Nhập lịch biểu",
-    "Photos" : "Ảnh",
     "Drive" : "Lái xe"
 },
 "nplurals=1; plural=0;");